refactor: 패키지 이동, 네이밍 정정, 테스트 별  명명
