+++ src/com/sh/views/InputVerification.java
@@ -1,4 +1,4 @@
-package com.sh.utils;
+package com.sh.views;
 
 import java.util.Arrays;
 import java.util.HashSet;
