+++ src/com/sh/views/Output.java
@@ -1,4 +1,4 @@
-package com.sh.utils;
+package com.sh.views;
 
 import java.util.function.Consumer;
