+++ test/com/sh/domains/PlayerTest.java
@@ -1,13 +1,14 @@
 package com.sh.domains;
 
-import static com.sh.utils.InputVerification.*;
+import static com.sh.views.InputVerification.*;
 import static org.junit.jupiter.api.Assertions.*;
 
-import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
 class PlayerTest {
+	@DisplayName("이름 - 예외")
 	@ParameterizedTest
 	@ValueSource(strings = {"abcdef", " "})
 	void invalid_name_of_player_exception(String name) {
@@ -16,6 +17,7 @@ class PlayerTest {
 		});
 	}
 
+	@DisplayName("정상 이름 - 플레이어 생성")
 	@ParameterizedTest
 	@ValueSource(strings = {"abcde", "honux", "JK"})
 	void acceptable_name_of_player_exception(String name) {
