+++ test/com/sh/domains/PlayersTest.java
@@ -1,14 +1,14 @@
 package com.sh.domains;
 
 import static com.sh.domains.LadderServiceTest.*;
-import static com.sh.utils.InputVerification.*;
+import static com.sh.views.InputVerification.*;
 import static java.util.stream.Collectors.*;
 import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.List;
 import java.util.stream.IntStream;
 
-
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 class PlayersTest {
@@ -16,6 +16,7 @@ class PlayersTest {
 	private List<String> invalidLongNames = getInvalidLongNames();
 	private int invalidLength = MAX_NUMBER_OF_PEOPLE + 1;
 
+	@DisplayName("적은 인원 - 예외")
 	@Test
 	void invalid_players_size_when_short() {
 		assertThrows(IllegalArgumentException.class, () -> {
@@ -23,6 +24,7 @@ class PlayersTest {
 		});
 	}
 
+	@DisplayName("한도 넘는 인원 - 예외")
 	@Test
 	void invalid_players_size_when_large() {
 		assertThrows(IllegalArgumentException.class, () -> {
@@ -30,6 +32,7 @@ class PlayersTest {
 		});
 	}
 
+	@DisplayName("플레이어스 생성")
 	@Test
 	void create_players() {
 		List<String> names = getNames();
